requirements.txt
ansible==2.9.22
ansible-lint==4.3.7
asn1crypto==0.24.0
attrs==19.3.0
bcrypt==3.1.7
cffi==1.12.3
colorama==0.4.4
commonmark==0.9.1
coverage==5.0.3
cryptography==2.7
importlib-metadata==1.5.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
more-itertools==8.2.0
netaddr==0.7.19
packaging==20.8
paramiko==2.6.0
passlib==1.7.1
pathspec==0.5.9
pluggy==0.13.1
py==1.8.1
pycparser==2.19
Pygments==2.7.3
PyNaCl==1.3.0
pyparsing==2.4.6
pytest==5.3.5
pytest-cov==2.8.1
PyYAML==5.1.1
rich==9.4.0
ruamel.yaml==0.16.0
ruamel.yaml.clib==0.1.0
six==1.12.0
typing-extensions==3.7.4.3
uuid==1.30
wcwidth==0.1.8
yamllint==1.25.0
zipp==3.0.0
